feat: scss removed, themes moved to themes.ts
