+++ react-ts-template/package.json
@@ -22,6 +22,7 @@
     "@mui/material": "^5.10.13",
     "D": "^1.0.0",
     "axios": "^1.1.3",
+    "clsx": "^1.2.1",
     "i18next": "^22.0.2",
     "jest": "^29.3.0",
     "mobx": "^6.6.2",
