--- react-ts-template/src/app.module.scss
@@ -1,5 +1,0 @@
-@import 'styles/index';
-
-.text {
-  color: $color-shades-red;
-}
