--- react-ts-template/src/styles/_index.scss
@@ -1,1 +1,0 @@
-@import 'variables';
