+++ react-ts-template/src/themes/index.ts
@@ -1,0 +1,3 @@
+import themes from './themes';
+
+export default themes;
